requirements.txt
astroid==1.6.4
autopep8==1.3.5
Babel==2.6.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.3.0
Django==2.0.3
django-appconf==1.0.2
django-compressor==2.2
django-libsass==0.7
django-modelcluster==4.1
django-phonenumber-field==2.0.0
django-pipeline==1.6.14
django-pipeline-typescript==0.2
django-taggit==0.22.2
django-treebeard==4.2.1
djangorestframework==3.7.7
draftjs-exporter==2.0.0
future==0.16.0
geocoder==1.38.1
geographiclib==1.49
geopy==1.15.0
html5lib==0.999999999
idna==2.6
isort==4.3.4
l18n==2016.6.4
lazy-object-proxy==1.3.1
libsass==0.14.5
mccabe==0.6.1
numpy==1.14.3
olefile==0.45.1
phonenumberslite==8.9.8
Pillow==4.3.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==1.9.1
pytz==2018.3
ratelim==0.1.6
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
wagtail==2.1.1
wagtail-nesting-box==0.2.1
wagtailgeowidget==4.0.3
wagtailgmaps==1.0.1
wagtailgridder==0.9.30
wagtailmenus==2.9
webencodings==0.5.1
Willow==1.1
wrapt==1.10.11
